requirements.txt
allure-behave
allure-pytest
selenium
pytest
pytest-bdd